Separar las rutas de consultas y hospitalizaciones
