+++ backend/app/api/main.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter
 
 from app.api.routes import beds, login, users, admins, \
-doctors, patients, documents, appointments
+doctors, patients, documents, consultations, hospitalizations
 
 api_router = APIRouter()
 api_router.include_router(beds.router, tags=["beds"])
@@ -11,4 +11,5 @@ api_router.include_router(admins.router, tags=["admins"])
 api_router.include_router(doctors.router, tags=["doctors"])
 api_router.include_router(patients.router, tags=["patients"])
 api_router.include_router(documents.router, tags=["documents"])
-api_router.include_router(appointments.router, tags=["appointments"])
+api_router.include_router(consultations.router, tags=["consultations"])
+api_router.include_router(hospitalizations.router, tags=["hospitalizations"])
