+++ backend/app/api/routes/__init__.py
@@ -5,4 +5,5 @@ import app.api.routes.admins as admins
 import app.api.routes.doctors as doctors
 import app.api.routes.patients as patients
 import app.api.routes.documents as documents
-import app.api.routes.appointments as appointments
+import app.api.routes.consultations as consultations
+import app.api.routes.hospitalizations as hospitalizations
